Refactor validations and dom ready / dom content loaded
- Validations are now handled in a separate file and made available to
the front-end through the data-form-validation attribute on the form
- Instead of dom ready / dom content loaded we now use the apostrophe
self.play functionality
