+++ lib/modules/user-form-widgets/index.js
@@ -1,5 +1,6 @@
 const rp          = require('request-promise');
 const styleSchema = require('../../../config/styleSchema.js').default;
+const mapFormValidations = require('./mapFormValidations');
 
 module.exports = {
     extend:    'apostrophe-widgets',
@@ -152,6 +153,7 @@ module.exports = {
         styleSchema.definition('containerStyles', 'Styles for the container')
     ],
     construct: function (self, options) {
+        
         const superLoad = self.load;
         self.load       = (req, widgets, callback) => {
             widgets.forEach((widget) => {
@@ -159,6 +161,7 @@ module.exports = {
                     const containerId               = styleSchema.generateId();
                     widget.containerId              = containerId;
                     widget.formattedContainerStyles = styleSchema.format(containerId, widget.containerStyles);
+                    widget.mappedValidation = mapFormValidations(widget.formFields);
                 }
                 
                 
@@ -173,6 +176,7 @@ module.exports = {
             superPushAssets();
             self.pushAsset('stylesheet', 'main', {when: 'always'});
             self.pushAsset('script', 'filepond', {when: 'always'});
+            self.pushAsset('script', 'init_filepond', {when: 'always'});
             self.pushAsset('script', 'main', {when: 'always'});
         };
